+++ tests/testthat/test_krona.R
@@ -10,12 +10,12 @@ if (class(krona_error_or_not) == "try-error"){
     expect_warning(merge_krona(c("Number.of.sequences.html", "Number.of.ASVs.html")))
   })
 } else {
-  test_that("krona unction works fine with GlobalPatterns dataset", {
+  test_that("krona function works fine with GlobalPatterns dataset", {
     testFolder <- tempdir()
     suppressWarnings(file.remove(list.files(testFolder, full.names = TRUE)))
-    expect_silent(krona(GA, paste0(testFolder, "/Number.of.sequences.html")))
-    XXX!!{ expect_silent(krona(GA, paste0(testFolder, "/Number.of.ASVs.html", nb_seq = F)))
-    XXX!!{ expect_silent(merge_krona(c(paste0(testFolder, "/Number.of.sequences.html"),
+    expect_silent(krona(GA, file = paste0(testFolder, "/Number.of.sequences.html")))
+    expect_silent(krona(GA, file = paste0(testFolder, "/Number.of.ASVs.html"), nb_seq = FALSE))
+    expect_silent(merge_krona(c(paste0(testFolder, "/Number.of.sequences.html"),
                                 paste0(testFolder, "/Number.of.ASVs.html"))))
   })
 }